:heavy_plus_sign: SuppliedValueStatistic.memoize()
