+++ src/test/java/org/terracotta/statistics/TableStatisticTest.java
@@ -31,7 +31,7 @@ import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertThat;
 import static org.terracotta.statistics.StatisticType.COUNTER;
 import static org.terracotta.statistics.StatisticType.GAUGE;
-import static org.terracotta.statistics.SuppliedValueStatistic.counter;
+import static org.terracotta.statistics.ValueStatistics.counter;
 
 public class TableStatisticTest {
